+++ lib/transpec/project.rb
@@ -1,5 +1,7 @@
 # coding: utf-8
 
+require 'transpec/rspec_version'
+
 module Transpec
   class Project
     attr_reader :path
@@ -15,6 +17,21 @@ module Transpec
     def require_bundler?
       gemfile_path = File.join(@path, 'Gemfile')
       File.exist?(gemfile_path)
+    end
+
+    def rspec_version
+      @rspec_version ||= begin
+        command = 'rspec --version'
+        command = 'bundle exec ' + command if require_bundler?
+
+        version_string = nil
+
+        Dir.chdir(@path) do
+          with_bundler_clean_env { version_string = `#{command}`.chomp }
+        end
+
+        RSpecVersion.new(version_string)
+      end
     end
 
     def with_bundler_clean_env