+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Support conversion to `be_truthy` / `be_falsey` ([#8](https://github.com/yujinakayama/transpec/issues/8))
 * Add `-b/--boolean-matcher` option that allows to specify matcher type that `be_true` and `be_false` will be converted to
+* Abort if a target project's `rspec` gem dependency is older than the version required by Transpec
 
 ## v1.1.2